[DSL] Move ordered memory operations to LLVM API
